Removed graphics layer, other misc fixes
