+++ include/api.h
@@ -42,7 +42,7 @@
 #define PROS_VERSION_MAJOR 3
 #define PROS_VERSION_MINOR 3
 #define PROS_VERSION_PATCH 2
-#define PROS_VERSION_STRING "3.3.2-dirty.46.8255ed1"
+#define PROS_VERSION_STRING "3.3.2-dirty.47.851466e"
 
 #define PROS_ERR (INT32_MAX)
 #define PROS_ERR_F (INFINITY)
